Use a custom lexer
The automatic lexers that come from lalrpop have a start up performance problem:
https://github.com/lalrpop/lalrpop/issues/1028
https://bugzilla.mozilla.org/show_bug.cgi?id=1978973
https://github.com/mozilla/jexl-rs/pull/32

This lexer is designed to match the semantics of the automatically generated lexer
exactly. That includes the weirdness around string escaping.

This improves the time to evaluate '1 + 2' from 370990 ns to 359 ns on my machine.

Note: This changes the API of ParseError to have a LexError instead of &str